LRCI-1352 Change around the class heirarchy to allow room for using Cucumber test results
