+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseTestResult.java
@@ -16,23 +16,14 @@ package com.liferay.jenkins.results.parser;
 
 import java.io.IOException;
 
-import java.net.MalformedURLException;
-import java.net.URISyntaxException;
-
 import java.util.Map;
 import java.util.Properties;
-
-import org.apache.commons.lang.StringUtils;
-
-import org.dom4j.Element;
-
-import org.json.JSONObject;
 
 /**
  * @author Leslie Wong
  * @author Yi-Chen Tsai
  */
-public class BaseTestResult implements TestResult {
+public abstract class BaseTestResult implements TestResult {
 
 	@Override
 	public Build getBuild() {
@@ -40,92 +31,66 @@ public class BaseTestResult implements TestResult {
 	}
 
 	@Override
-	public String getClassName() {
-		return _className;
-	}
+	public boolean isFailing() {
+		String status = getStatus();
 
-	@Override
-	public String getDisplayName() {
-		String testName = getTestName();
+		if (status.equals("FIXED") || status.equals("PASSED") ||
+			status.equals("SKIPPED")) {
 
-		if (testName.startsWith("test[")) {
-			return testName.substring(5, testName.length() - 1);
+			return false;
 		}
 
-		return getSimpleClassName() + "." + testName;
+		return true;
 	}
 
 	@Override
-	public long getDuration() {
-		return _duration;
+	public boolean isUniqueFailure() {
+		return !UpstreamFailureUtil.isTestFailingInUpstreamJob(this);
 	}
 
-	@Override
-	public String getErrorDetails() {
-		return _errorDetails;
-	}
-
-	@Override
-	public String getErrorStackTrace() {
-		return _errorStackTrace;
-	}
-
-	@Override
-	public Element getGitHubElement() {
-		String testReportURL = getTestReportURL();
-
-		Element downstreamBuildListItemElement = Dom4JUtil.getNewElement(
-			"div", null);
-
-		downstreamBuildListItemElement.add(
-			Dom4JUtil.getNewAnchorElement(testReportURL, getDisplayName()));
-
-		String errorStackTrace = getErrorStackTrace();
-
-		if ((errorStackTrace != null) && !errorStackTrace.isEmpty()) {
-			String trimmedStackTrace = StringUtils.abbreviate(
-				errorStackTrace, _LINES_ERROR_STACK_DISPLAY_SIZE_MAX);
-
-			downstreamBuildListItemElement.add(
-				Dom4JUtil.toCodeSnippetElement(trimmedStackTrace));
+	protected BaseTestResult(Build build) {
+		if (build == null) {
+			throw new IllegalArgumentException("Build is null");
 		}
 
-		return downstreamBuildListItemElement;
+		_build = build;
 	}
 
-	@Override
-	public String getPackageName() {
-		String className = getClassName();
+	protected String getAxisNumber() {
+		Build build = getBuild();
 
-		int x = className.lastIndexOf(".");
+		if (build instanceof AxisBuild) {
+			AxisBuild axisBuild = (AxisBuild)build;
 
-		if (x < 0) {
-			return "(root)";
+			return axisBuild.getAxisNumber();
 		}
 
-		return className.substring(0, x);
+		return "INVALID_AXIS_NUMBER";
 	}
 
-	@Override
-	public String getSimpleClassName() {
-		String className = getClassName();
+	protected String getConsoleOutputURL() {
+		StringBuilder sb = new StringBuilder();
 
-		int x = className.lastIndexOf(".");
+		sb.append(getTestrayLogsURL());
+		sb.append("/jenkins-console.txt.gz");
 
-		return className.substring(x + 1);
+		return sb.toString();
 	}
 
-	@Override
-	public String getStatus() {
-		return _status;
+	protected String getLiferayLogURL() {
+		StringBuilder sb = new StringBuilder();
+
+		String name = getDisplayName();
+
+		sb.append(getTestrayLogsURL());
+		sb.append("/");
+		sb.append(name.replace('#', '_'));
+		sb.append("/liferay-log.txt.gz");
+
+		return sb.toString();
 	}
 
-	@Override
-	public String getTestName() {
-		return _testName;
-	}
-
-	public String getTestrayLogsURL() {
+	protected String getTestrayLogsURL() {
 		Properties buildProperties = null;
 
 		try {
@@ -160,112 +125,23 @@ public class BaseTestResult implements TestResult {
 			build.getJobVariant(), "/", getAxisNumber());
 	}
 
-	@Override
-	public String getTestReportURL() {
-		StringBuilder sb = new StringBuilder();
+	protected boolean hasLiferayLog() {
+		String liferayLog = null;
 
-		Build build = getBuild();
-
-		sb.append(build.getBuildURL());
-
-		sb.append("/testReport/");
-		sb.append(getPackageName());
-		sb.append("/");
-		sb.append(getSimpleClassName());
-		sb.append("/");
-		sb.append(getEncodedTestName());
-
-		String testReportURL = sb.toString();
-
-		if (testReportURL.startsWith("http")) {
-			try {
-				return JenkinsResultsParserUtil.encode(testReportURL);
-			}
-			catch (MalformedURLException | URISyntaxException exception) {
-				System.out.println(
-					"Unable to encode the test report " + testReportURL);
-			}
+		try {
+			liferayLog = JenkinsResultsParserUtil.toString(
+				getLiferayLogURL(), false, 0, 0, 0);
 		}
-
-		return testReportURL;
-	}
-
-	@Override
-	public boolean isFailing() {
-		if (_status.equals("FIXED") || _status.equals("PASSED") ||
-			_status.equals("SKIPPED")) {
-
+		catch (IOException ioException) {
 			return false;
 		}
 
-		return true;
+		return !liferayLog.isEmpty();
 	}
-
-	@Override
-	public boolean isUniqueFailure() {
-		return !UpstreamFailureUtil.isTestFailingInUpstreamJob(this);
-	}
-
-	protected BaseTestResult(Build build, JSONObject caseJSONObject) {
-		if (build == null) {
-			throw new IllegalArgumentException("Build is null");
-		}
-
-		_build = build;
-		_className = caseJSONObject.getString("className");
-		_duration = (long)(caseJSONObject.getDouble("duration") * 1000D);
-		_status = caseJSONObject.getString("status");
-		_testName = caseJSONObject.getString("name");
-
-		if (_status.equals("FAILED") && caseJSONObject.has("errorDetails") &&
-			caseJSONObject.has("errorStackTrace")) {
-
-			_errorDetails = caseJSONObject.optString("errorDetails");
-			_errorStackTrace = caseJSONObject.optString("errorStackTrace");
-		}
-		else {
-			_errorDetails = null;
-			_errorStackTrace = null;
-		}
-	}
-
-	protected String getAxisNumber() {
-		Build build = getBuild();
-
-		if (build instanceof AxisBuild) {
-			AxisBuild axisBuild = (AxisBuild)build;
-
-			return axisBuild.getAxisNumber();
-		}
-
-		return "INVALID_AXIS_NUMBER";
-	}
-
-	protected String getEncodedTestName() {
-		StringBuilder sb = new StringBuilder(getTestName());
-
-		for (int i = 0; i < sb.length(); i++) {
-			char c = sb.charAt(i);
-
-			if (!Character.isJavaIdentifierPart(c)) {
-				sb.setCharAt(i, '_');
-			}
-		}
-
-		return sb.toString();
-	}
-
-	private static final int _LINES_ERROR_STACK_DISPLAY_SIZE_MAX = 1500;
 
 	private static final String _URL_BASE_LOGS_DEFAULT =
 		"https://testray.liferay.com/reports/production/logs";
 
 	private final Build _build;
-	private final String _className;
-	private final long _duration;
-	private final String _errorDetails;
-	private final String _errorStackTrace;
-	private final String _status;
-	private final String _testName;
 
 }
